+++ tests/test_palette.py
@@ -25,31 +25,11 @@ def test_02():  # image size with 16 swatches
     assert img.size == (800, 800)
 
 
-def test_03():  # image size with two columns
-    img = palette.draw_swatches(colors=COLORS, cols=2)
-    assert img.size == (400, 800)
-
-
-def test_04():  # image size with 1 row
-    img = palette.draw_swatches(colors=COLORS, cols=8)
-    assert img.size == (1600, 200)
-
-
-def test_05():  # image size with too many columns
-    img = palette.draw_swatches(colors=COLORS, cols=80)
-    assert img.size == (1600, 200)
-
-
-def test_06():  # image size with 100 px swatcher
+def test_03():  # image size with 100 px swatcher
     img = palette.draw_swatches(colors=COLORS, size=100)
     assert img.size == (600, 300)
 
 
-def test_07():  # image size with less colors than cols
+def test_04():  # image size with less colors than cols
     img = palette.draw_swatches(COLORS[:2])
     assert img.size == (400, 200)
-
-
-def test_08():  # image size with lots of colors
-    img = palette.draw_swatches(colors=COLORS * 10, cols=10)
-    assert img.size == (2000, 1600)